# contracts/openapi-v2.yaml
openapi: 3.1.0
info:
  title: Woorisai API
  version: 2.0.0
  description: >-
    Private API for exactly two participants. Protected operations use stateless
    HTTP Basic credentials: the username is participant slot `1` or `2`, and the
    password is exactly four ASCII digits. Credentials are verified on every request.
...
security:
  - BasicAuth: []
paths:
...
  /api/v2/relationship-scores:
    get:
      tags:
        - Relationship
      summary: Read the authenticated participant's outgoing and incoming scores
      operationId: getRelationshipScores
      responses:
        "200":
          description: The canonical participant pair and both directional scores
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipScoresResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/RelationshipForbidden"
        "503":
          $ref: "#/components/responses/RelationshipOrAuthenticationUnavailable"
  /api/v2/score-changes:
    get:
      tags:
        - Relationship
      summary: List score changes from both directions, newest first
      operationId: listScoreChanges
      parameters:
        - $ref: "#/components/parameters/PageNumber"
      responses:
        "200":
          description: One-based page of score changes
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreChangeHistoryResponse"
        "400":
          $ref: "#/components/responses/InvalidRelationshipRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/RelationshipForbidden"
        "404":
          $ref: "#/components/responses/RelationshipNotFound"
        "503":
          $ref: "#/components/responses/RelationshipOrAuthenticationUnavailable"
    post:
      tags:
        - Relationship
      summary: Change only the authenticated participant's outgoing score
      operationId: createScoreChange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeScoreRequest"
      responses:
        "201":
          description: The immutable change and new outgoing score
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreChangeCreatedResponse"
        "400":
          $ref: "#/components/responses/InvalidRelationshipRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/RelationshipForbidden"
        "404":
          $ref: "#/components/responses/RelationshipNotFound"
        "409":
          $ref: "#/components/responses/RelationshipConflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          $ref: "#/components/responses/RelationshipOrAuthenticationUnavailable"
  /api/v2/score-changes/{scoreChangeId}:
    parameters:
      - $ref: "#/components/parameters/ScoreChangeId"
    get:
      tags:
        - Relationship
      summary: Read one score change and its flat chronological comment thread
      operationId: getScoreChange
      responses:
        "200":
          description: The score change and comments in ascending creation order
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreChangeThreadResponse"
        "400":
          $ref: "#/components/responses/InvalidRelationshipRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/RelationshipForbidden"
        "404":
          $ref: "#/components/responses/RelationshipNotFound"
        "503":
          $ref: "#/components/responses/RelationshipOrAuthenticationUnavailable"
  /api/v2/score-changes/{scoreChangeId}/comments:
    parameters:
      - $ref: "#/components/parameters/ScoreChangeId"
    post:
      tags:
        - Relationship
      summary: Add a text, media, or text-and-media comment to a score change
      operationId: createScoreChangeComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoreChangeCommentRequest"
      responses:
        "201":
          description: The created comment
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreChangeCommentCreatedResponse"
        "400":
          $ref: "#/components/responses/InvalidRelationshipRequest"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/RelationshipForbidden"
        "404":
          $ref: "#/components/responses/RelationshipNotFound"
        "409":
          $ref: "#/components/responses/RelationshipConflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          $ref: "#/components/responses/RelationshipOrAuthenticationUnavailable"
...
  /api/v2/notification-fids:
    post:
      tags:
        - Notification
      summary: Register or globally reassign a Firebase installation ID
      operationId: registerNotificationFid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationFidRequest"
      responses:
        "204":
          description: The FID now belongs to the authenticated participant
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "400":
          $ref: "#/components/responses/InvalidNotificationFid"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          $ref: "#/components/responses/NotificationFidOrAuthenticationUnavailable"
    delete:
      tags:
        - Notification
      summary: Unregister a FID only when owned by the authenticated participant
      operationId: unregisterNotificationFid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationFidRequest"
      responses:
        "204":
          description: The owned FID was removed; missing or foreign FIDs are no-ops
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
        "400":
          $ref: "#/components/responses/InvalidNotificationFid"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          $ref: "#/components/responses/NotificationFidOrAuthenticationUnavailable"
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: >-
        Send `Authorization: Basic base64("<slot>:<PIN>")`. The username is
        exactly `1` or `2`; the password matches `^[0-9]{4}$` using ASCII digits.
  headers:
    NoStore:
      description: The response Cache-Control value contains the `no-store` directive
      required: true
      schema:
        type: string
        pattern: '(^|.*,[ ]*)no-store([ ]*,.*|$)'
    BasicChallenge:
      description: Requests participant slot and PIN HTTP Basic credentials
      required: true
      schema:
        type: string
        enum:
          - 'Basic realm="woorisai"'
  parameters:
    PageNumber:
      name: pageNumber
      in: query
      required: false
      description: One-based page number; omitted values default to 1
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 2147483647
        default: 1
...
    ScoreChangeId:
      name: scoreChangeId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/PositiveInt64"
...
  responses:
    AuthenticationRequired:
      description: Basic credentials are missing, malformed, or do not match
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        WWW-Authenticate:
          $ref: "#/components/headers/BasicChallenge"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/AuthenticationRequiredProblem"
    UnsupportedMediaType:
      description: A JSON request body was sent with a Content-Type other than application/json
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/UnsupportedMediaTypeProblem"
...
    InvalidRelationshipRequest:
      description: The relationship request is malformed or violates input limits
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InvalidRelationshipRequestProblem"
    RelationshipNotFound:
      description: The score change or requested media upload was not found
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/RelationshipNotFoundProblem"
    RelationshipForbidden:
      description: The relationship resource or media upload is not owned by this participant
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/RelationshipForbiddenProblem"
    RelationshipConflict:
      description: The score or attachment request conflicts with current state
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/RelationshipConflictProblem"
    RelationshipOrAuthenticationUnavailable:
      description: Authentication, relationship, participant, media, or persistence data is unavailable
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/AuthenticationUnavailableProblem"
              - $ref: "#/components/schemas/RelationshipUnavailableProblem"
...
    InvalidNotificationFid:
      description: The request does not contain one exact 22-character base64url FID
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InvalidNotificationFidProblem"
    NotificationFidOrAuthenticationUnavailable:
      description: Authentication or notification FID persistence is unavailable
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/AuthenticationUnavailableProblem"
              - $ref: "#/components/schemas/NotificationFidUnavailableProblem"
  schemas:
    CanonicalUuid:
      type: string
      format: uuid
      pattern: '^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$'
    PositiveInt64:
      type: integer
      format: int64
      minimum: 1
...
    StoredMediaContentType:
      type: string
      enum:
        - image/jpeg
        - image/png
        - image/webp
        - video/mp4
        - video/webm
        - video/quicktime
    AttachedMedia:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - fileName
        - contentType
        - byteSize
      properties:
        id:
          $ref: "#/components/schemas/CanonicalUuid"
        kind:
          type: string
          enum:
            - IMAGE
            - VIDEO
        fileName:
          type: string
          minLength: 1
          maxLength: 255
        contentType:
          $ref: "#/components/schemas/StoredMediaContentType"
        byteSize:
          type: integer
          format: int64
          minimum: 1
          maximum: 104857600
    FlexibleAttachmentGroup:
      type: array
      maxItems: 4
      description: Empty, at most four images, or exactly one video; image/video mixing is invalid
      items:
        $ref: "#/components/schemas/AttachedMedia"
      oneOf:
        - type: array
          maxItems: 4
          items:
            allOf:
              - $ref: "#/components/schemas/AttachedMedia"
              - type: object
                properties:
                  kind:
                    enum:
                      - IMAGE
                  contentType:
                    enum:
                      - image/jpeg
                      - image/png
                      - image/webp
                  byteSize:
                    maximum: 10485760
        - type: array
          minItems: 1
          maxItems: 1
          items:
            allOf:
              - $ref: "#/components/schemas/AttachedMedia"
              - type: object
                properties:
                  kind:
                    enum:
                      - VIDEO
                  contentType:
                    enum:
                      - video/mp4
                      - video/webm
                      - video/quicktime
    ScoreChangeAttachedMedia:
      allOf:
        - $ref: "#/components/schemas/AttachedMedia"
        - type: object
          properties:
            kind:
              enum:
                - IMAGE
            contentType:
              enum:
                - image/jpeg
                - image/png
                - image/webp
            byteSize:
              maximum: 10485760
    RelationshipParticipant:
      type: object
      additionalProperties: false
      required:
        - slot
        - displayName
        - mine
      properties:
        slot:
          type: integer
          format: int32
          enum:
            - 1
            - 2
        displayName:
          type: string
          minLength: 1
          maxLength: 30
        mine:
          type: boolean
    RelationshipScore:
      type: object
      additionalProperties: false
      required:
        - sourceParticipant
        - targetParticipant
        - currentScore
        - updatedAt
      properties:
        sourceParticipant:
          $ref: "#/components/schemas/RelationshipParticipant"
        targetParticipant:
          $ref: "#/components/schemas/RelationshipParticipant"
        currentScore:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
        updatedAt:
          type: string
          format: date-time
    RelationshipScoresResponse:
      type: object
      additionalProperties: false
      required:
        - self
        - partner
        - outgoing
        - incoming
      properties:
        self:
          $ref: "#/components/schemas/RelationshipParticipant"
        partner:
          $ref: "#/components/schemas/RelationshipParticipant"
        outgoing:
          $ref: "#/components/schemas/RelationshipScore"
        incoming:
          $ref: "#/components/schemas/RelationshipScore"
    ChangeScoreRequest:
      type: object
      additionalProperties: false
      description: >-
        Exactly one of delta or targetScore must be present. A supplied delta must be
        non-zero. These cross-field rules are enforced by the server and return
        INVALID_RELATIONSHIP_REQUEST when violated.
      properties:
        delta:
          type:
            - integer
            - "null"
          format: int32
          minimum: -100
          maximum: 100
          description: A non-null value must be non-zero
        targetScore:
          type:
            - integer
            - "null"
          format: int32
          minimum: 0
          maximum: 100
        reason:
          type:
            - string
            - "null"
          maxLength: 200
          description: Blank values are normalized to null; omission means no reason
        mediaUploadIds:
          type:
            - array
            - "null"
          maxItems: 1
          uniqueItems: true
          description: >-
            Completed, parentless SCORE_CHANGE uploads owned by the authenticated
            participant. The final group is empty or exactly one image; video is rejected.
          items:
            $ref: "#/components/schemas/CanonicalUuid"
    ScoreChange:
      type: object
      additionalProperties: false
      required:
        - id
        - sourceParticipant
        - targetParticipant
        - changedBy
        - delta
        - resultingScore
        - reason
        - createdAt
        - commentCount
        - attachments
      properties:
        id:
          $ref: "#/components/schemas/PositiveInt64"
        sourceParticipant:
          $ref: "#/components/schemas/RelationshipParticipant"
        targetParticipant:
          $ref: "#/components/schemas/RelationshipParticipant"
        changedBy:
          $ref: "#/components/schemas/RelationshipParticipant"
        delta:
          type: integer
          format: int32
          minimum: -100
          maximum: 100
          description: Always non-zero; enforced by the server and persisted domain model
        resultingScore:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
        reason:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 200
        createdAt:
          type: string
          format: date-time
        commentCount:
          type: integer
          format: int64
          minimum: 0
        attachments:
          type: array
          maxItems: 1
          items:
            $ref: "#/components/schemas/ScoreChangeAttachedMedia"
    ScoreChangeComment:
      type: object
      additionalProperties: false
      required:
        - id
        - author
        - content
        - createdAt
        - attachments
      properties:
        id:
          $ref: "#/components/schemas/PositiveInt64"
        author:
          $ref: "#/components/schemas/RelationshipParticipant"
        content:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 500
          description: Null only for a media-only comment
        createdAt:
          type: string
          format: date-time
        attachments:
          $ref: "#/components/schemas/FlexibleAttachmentGroup"
      description: >-
        Content is null only for a media-only comment. The server guarantees that
        content is non-null or attachments is non-empty.
    ScoreChangeHistoryResponse:
      type: object
      additionalProperties: false
      required:
        - results
        - paging
      properties:
        results:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ScoreChange"
        paging:
          $ref: "#/components/schemas/RelationshipPaging"
    RelationshipPaging:
      type: object
      additionalProperties: false
      required:
        - pageNumber
        - pageSize
        - hasNext
        - totalCount
      properties:
        pageNumber:
          type: integer
          format: int32
          minimum: 1
        pageSize:
          type: integer
          format: int32
          enum:
            - 20
        hasNext:
          type: boolean
        totalCount:
          type: integer
          format: int64
          minimum: 0
    ScoreChangeCreatedResponse:
      type: object
      additionalProperties: false
      required:
        - change
        - outgoing
      properties:
        change:
          $ref: "#/components/schemas/ScoreChange"
        outgoing:
          $ref: "#/components/schemas/RelationshipScore"
    ScoreChangeThreadResponse:
      type: object
      additionalProperties: false
      required:
        - change
        - comments
      properties:
        change:
          $ref: "#/components/schemas/ScoreChange"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/ScoreChangeComment"
    CreateScoreChangeCommentRequest:
      type: object
      additionalProperties: false
      description: >-
        At least normalized content or one mediaUploadId must be supplied. The server
        enforces this cross-field rule and returns INVALID_RELATIONSHIP_REQUEST when
        both are absent or empty.
      properties:
        content:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 500
          description: Values empty after server-side surrounding whitespace stripping are normalized to null
        mediaUploadIds:
          type:
            - array
            - "null"
          maxItems: 4
          uniqueItems: true
          description: >-
            Completed, parentless SCORE_CHANGE_COMMENT uploads owned by the authenticated
            participant. The final group is empty, up to four images, or exactly one video;
            images and video cannot be mixed.
          items:
            $ref: "#/components/schemas/CanonicalUuid"
    ScoreChangeCommentCreatedResponse:
      type: object
      additionalProperties: false
      required:
        - comment
      properties:
        comment:
          $ref: "#/components/schemas/ScoreChangeComment"
...
    NotificationFidRequest:
      type: object
      additionalProperties: false
      required:
        - fid
      properties:
        fid:
          type: string
          minLength: 22
          maxLength: 22
          pattern: '^[A-Za-z0-9_-]{22}$'
    ApiProblem:
      type: object
      additionalProperties: false
      required:
        - title
        - status
        - detail
        - instance
        - errorCode
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        errorCode:
          type: string
    NotificationApiProblem:
      type: object
      additionalProperties: false
      required:
        - title
        - status
        - detail
        - errorCode
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        errorCode:
          type: string
    AuthenticationRequiredProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Authentication required]
            status:
              enum: [401]
            detail:
              enum: [Valid HTTP Basic participant credentials are required.]
            errorCode:
              enum: [AUTHENTICATION_REQUIRED]
    AuthenticationUnavailableProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Authentication unavailable]
            status:
              enum: [503]
            detail:
              enum: [Authentication is temporarily unavailable.]
            errorCode:
              enum: [AUTHENTICATION_UNAVAILABLE]
    UnsupportedMediaTypeProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Unsupported media type]
            status:
              enum: [415]
            detail:
              enum: [Content-Type must be application/json.]
            errorCode:
              enum: [UNSUPPORTED_MEDIA_TYPE]
...
    InvalidRelationshipRequestProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Invalid relationship request]
            status:
              enum: [400]
            detail:
              enum: [The relationship request is invalid.]
            errorCode:
              enum: [INVALID_RELATIONSHIP_REQUEST]
    RelationshipNotFoundProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Relationship resource not found]
            status:
              enum: [404]
            detail:
              enum: [The requested relationship resource was not found.]
            errorCode:
              enum: [RELATIONSHIP_NOT_FOUND]
    RelationshipForbiddenProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Relationship access denied]
            status:
              enum: [403]
            detail:
              enum: [Access to this relationship resource is denied.]
            errorCode:
              enum: [RELATIONSHIP_FORBIDDEN]
    RelationshipConflictProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Relationship conflict]
            status:
              enum: [409]
            detail:
              enum: [The relationship request conflicts with current state.]
            errorCode:
              enum: [RELATIONSHIP_CONFLICT]
    RelationshipUnavailableProblem:
      allOf:
        - $ref: "#/components/schemas/ApiProblem"
        - type: object
          properties:
            title:
              enum: [Relationship unavailable]
            status:
              enum: [503]
            detail:
              enum: [Relationship data is temporarily unavailable.]
            errorCode:
              enum: [RELATIONSHIP_UNAVAILABLE]
...
    InvalidNotificationFidProblem:
      allOf:
        - $ref: "#/components/schemas/NotificationApiProblem"
        - type: object
          properties:
            title:
              enum: [Invalid notification FID request]
            status:
              enum: [400]
            detail:
              enum: [Request must contain one valid Firebase installation ID.]
            errorCode:
              enum: [INVALID_NOTIFICATION_FID]
    NotificationFidUnavailableProblem:
      allOf:
        - $ref: "#/components/schemas/NotificationApiProblem"
        - type: object
          properties:
            title:
              enum: [Notification FID service unavailable]
            status:
              enum: [503]
            detail:
              enum: [Notification FID service is temporarily unavailable.]
            errorCode:
              enum: [NOTIFICATION_FID_UNAVAILABLE]